paradex_jwt_auth/requirements.txt
python-dotenv
paradex_py
requests
